move mesh point selector implementation to examples
