+++ genesis/vis/viewer_plugins/__init__.py
@@ -2,5 +2,6 @@ from .plugins import *
 from .viewer_plugin import (
     EVENT_HANDLE_STATE,
     EVENT_HANDLED,
+    RaycasterPlugin,
     ViewerPlugin,
 )
