+++ genesis/vis/viewer_plugins/plugins/__init__.py
@@ -1,9 +1,7 @@
 from .default_controls import DefaultControlsPlugin
-from .mesh_point_selector import MeshPointSelectorPlugin
 from .mouse_interaction import MouseInteractionPlugin
 
 __all__ = [
     "DefaultControlsPlugin",
-    "MeshPointSelectorPlugin",
     "MouseInteractionPlugin",
 ]
